provide --save-temps option keep generated C source and header files when
2007-08-08 Mathias Hasselmann <mathias.hasselmann@gmx.de>

	* compiler/valacompiler.vala: provide --save-temps option
	* gobject/valaccodecompiler.vala: keep generated C source and header
	files when --save-temps, respectively --library options are used
	* vala/valacodecontext.vala: add save_temps, save_csources
	and save_cheaders property

svn path=/trunk/; revision=452

+++ gobject/valaccodecompiler.vala
@@ -107,8 +107,10 @@ public class Vala.CCodeCompiler {
 		/* remove generated C source and header files */
 		foreach (SourceFile file in source_files) {
 			if (!file.pkg) {
-				FileUtils.unlink (file.get_csource_filename ());
-				FileUtils.unlink (file.get_cheader_filename ());
+				if (!context.save_csources)
+					FileUtils.unlink (file.get_csource_filename ());
+				if (!context.save_cheaders)
+					FileUtils.unlink (file.get_cheader_filename ());
 			}
 		}
 	}
